refactor(contract): move node migration into api/node_migration.rs
