+++ crates/contract/src/api.rs
@@ -7,6 +7,7 @@ pub mod governance;
 pub mod key_events;
 pub mod keys;
 pub mod lifecycle;
+pub mod node_migration;
 pub mod sign;
 #[cfg(not(target_arch = "wasm32"))]
 #[cfg(test)]
